melib/logging: manually redact http auth info
We don't log auth info on IMAP because we handle the connection
implementation. But in JMAP the underlying connection is an HTTP
connection from the isahc crate, which logs on its own. If the log level
is DEBUG or TRACE, this can result in Bearer tokens being leaked in the
logs.

This commit makes a crude check and replace with a regular expression
for Basic and Bearer HTTP auth scheme logs.

Fixes #566 ("Redact authentication credentials from http (isahc) logs")

Resolves: https://git.meli-email.org/meli/meli/issues/566
Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>