melib/imap: only process parsed untagged responses
Instead of trying to parse an untagged response inside the
`process_untagged` method, pass only parsed untagged responses to it,
and return any refresh events as an `Option`.

This refactoring allows the `process_untagged` called to decide what
to do with any potential backend refresh events: if the caller is a
watch future (IDLE/polling) it can return them as part of the watch
`Stream` async generator, otherwise it can forward them to the
connection's backend event consumer to forward it to the front-end.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>