melib/error: add TLSConnectionFailed kind variant
Commit 919fd00c ("melib/error: hardcode native-tls error check")
introduced a hard coded check for errors from the `openssl` crate in
order to enable recovery from terminated connections for the NNTP
backend. However, the InvalidTLSConnection variant it used was not
judged as "recoverable" in the `ErrorKind::is_recoverable` because
`NetworkErrorKind::is_network_down` returned false for that variant.

This commit introduced a new `NetworkErrorKind` variant,
`TLSConnectionFailed` that is recoverable.

Fixes: 919fd00c ("melib/error: hardcode native-tls error check")
Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>