melib/notmuch: split symbol name in call!() macro
call!() macro receives a type as an argument for which function to call
from the dynamic library. This type is used for typechecking the
returned symbol from libloading. The type argument is also used to
compute the symbol name as bytes to pass to libloading. This meant we
cannot pass type paths such as ffi::SYMBOL to this macro, only SYMBOL.

Rewrite macro to split the stringified type by "::" instead.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>