melib/notmuch: rewrite watching logic
Rewrite watching logic to poll both every few seconds and whenever a
filesystem event is received with `notify` watcher.

Remove reliance on database revision because it proves to be unreliable
and can only be used for detecting added emails with a query, and
nothing else.

This commit adds a `Snapshot` struct that stores the latest database
connection.

When trying to detect new events, query the snapshot's database files
and compare them to a new connection's files; compute any differences
from that and if any changes were detected, store the new connection to
the snapshot.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>