melib/mbox: check for CRLF when parsing
When parsing mbox files, we made the mistake to assume unix line endings
(\n). This could lead to panics because of out of bounds access or
otherwise ignored messages due to invalid parsing.

This commit changes all parsing functions to optionally use CRLF if the
byte slice contains it.

Fixes #585 ("Thread panic when reading an mbox file")

Resolves: https://git.meli-email.org/meli/meli/issues/585
Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>