melib/jmap: Fix SetResponse struct definition
The Set response object fields `notCreated`, `notUpdated, `notDestroyed`
are all maps of `SetError`s indexed by object `Id`, not
sequences/vectors.

If a Set call failed to alter some objects, the Set response would not
be deserialized and a deserialization error would be returned.

This commit makes the Vecs into IndexMaps and furthermore adjust the
code locations where a SetResponse is inspected to process the fields
properly.

Fixes #593 ("jmap response cannot be decoded")

Resolves: https://git.meli-email.org/meli/meli/issues/593
Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>